August 2014 - Version 1.3
This round of changes should combine C. Price's edits to the Python
files and the Arc Toolbox, and D. Ignizio's updates to the entire
package, complete with added features. This single version of the
toolbox should work in ESRI 10.0, 10.1, and 10.2. MP can now be run from
within the tool in a new tab.