Bug fix for Vertical Coordinate Reference Systems
This fix should correct the issue discovered by J. Krstolic in regard to
the improperly constructed (or failed) Spatial Reference section that
presented itself as a result of VCRS information being present. I'm
unsure as to whether the VCRS information was in her template or in the
dataset itself, but the updates resolved the issue when testing against
the data she sent.